p4i integration Integrate change 72436 from 3.0 to main. Description for change 72436:
: Change 72436 by vohi@vohi-hokkaido-nt on 2002/09/02 10:18:41
: 
: 	Compile without warning.
: 
: Affected files ...
: 
: ... //depot/qt/3.0/src/widgets/qpopupmenu.cpp#53 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 72437]

+++ src/widgets/qpopupmenu.cpp
@@ -1878,7 +1878,7 @@ void QPopupMenu::keyPressEvent( QKeyEvent *e )
 	    QString s = m->text();
 	    if ( !s.isEmpty() ) {
 		int i = s.find( '&' );
-		while ( i >= 0 && i < s.length() - 1 ) {
+		while ( i >= 0 && i < (int)s.length() - 1 ) {
 		    if ( s[i+1].upper() == c ) {
 			ok_key = TRUE;
 			clashCount++;
